Remove the config loader hooks that are unnecessary
