+++ app/projects/query.tsx
@@ -1,6 +1,3 @@
-import { useRouteLoaderData } from '@remix-run/react'
-import type { QueryConfigLoader } from '~/routes/query.$version.docs'
-
 export const gradientText =
   'inline-block text-transparent bg-clip-text bg-gradient-to-r from-red-500 to-amber-500'
 
@@ -38,15 +35,3 @@ export function getBranch(argVersion?: string) {
 }
 
 export type Framework = 'angular' | 'react' | 'svelte' | 'vue' | 'solid'
-
-export const useQueryDocsConfig = () => {
-  const queryConfigLoaderData = useRouteLoaderData<QueryConfigLoader>(
-    'routes/query.$version.docs'
-  )
-
-  if (!queryConfigLoaderData?.tanstackDocsConfig) {
-    throw new Error('Config could not be read for tanstack/query!')
-  }
-
-  return queryConfigLoaderData
-}
