+++ app/routes/ranger.v1.docs.tsx
@@ -1,15 +1,9 @@
 import * as React from 'react'
 import { FaDiscord, FaGithub } from 'react-icons/fa'
-import { Link, json } from '@remix-run/react'
+import { Link, json, useLoaderData } from '@remix-run/react'
 import type { MetaFunction } from '@remix-run/node'
-import {
-  gradientText,
-  repo,
-  useRangerV1Config,
-  v1branch,
-} from '~/projects/ranger'
+import { gradientText, repo, v1branch } from '~/projects/ranger'
 import { seo } from '~/utils/seo'
-import type { DocsConfig } from '~/components/Docs'
 import { Docs } from '~/components/Docs'
 import { getTanstackDocsConfig } from '~/utils/config'
 
@@ -20,8 +14,6 @@ export const loader = async () => {
     tanstackDocsConfig,
   })
 }
-
-export type RangerConfigV1Loader = typeof loader
 
 const logo = (
   <>
@@ -69,15 +61,14 @@ export const meta: MetaFunction = () => {
 }
 
 export default function DocsRoute() {
-  let config = useRangerV1Config()
+  const { tanstackDocsConfig } = useLoaderData<typeof loader>()
 
-  config = React.useMemo(
-    () =>
-      ({
-        ...config,
-        menu: [localMenu, ...config.menu],
-      } as DocsConfig),
-    [config]
+  const config = React.useMemo(
+    () => ({
+      ...tanstackDocsConfig,
+      menu: [localMenu, ...tanstackDocsConfig.menu],
+    }),
+    [tanstackDocsConfig]
   )
 
   return (
